requirements.txt
PyQt5
mysql-connector-python
pyinstaller
